Add -n option, fix indenting, add comments
This commit turned into a bit of a hodge-podge, so bear with me.

First, the -n option was added. I don't think I'll ever use this on my
own, as I quite like the formatted indentation of the output, but I can
imagine some others might not like it as much.

I also changed the order of options slightly in usage() and README.md.
Options are still alphabetized, but when listed in documentation, all
options without arguments are listed, and then all options with
arguments.

In the midst of this I also fixed some incorrect indentation which has
been bugging me. For whatever reason, multi-line statements don't work well
with Neovim indenting (or really, Bash itself, for that matter), so I put
them all on one line. After the complicated pipe statements, which
include several calls to sed, I put block comments explaining the
logic of the section. I'm slightly annoyed that I now have a few lines
which are extra long, but I don't know what else to do. Oh well.