Fix printf bug
This fixes a bug where the videos for a particular channel will not be
listed if any video had a '%' in the title. I first noticed this bug
when testing for the -n option, and I saw this output:

./omnavi: line 69: printf: `S': invalid format character

Upon investigation, a video in my list had a '%' followed by an 'S'.

This is easily fixed with the sed statement 's/%/%%/g' as seen below.